apifinal/myapi/requirements.txt
async-generator==1.10
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
python-dateutil==2.8.2
requests==2.28.2

